refactor: Make collections 2.13 cross-compatible
