+++ cognitive/src/test/scala/com/microsoft/azure/synapse/ml/cognitive/form/FormRecognizerV3Suite.scala
@@ -32,7 +32,7 @@ object FormRecognizerV3Utils {
 
   def resultAssert(result: Array[Row], str1: String, str2: String): Unit = {
     assert(result.head.getString(2).startsWith(str1))
-    assert(result.head.getSeq(3).head.asInstanceOf[HashMap.HashTrieMap[String, _]]
+    assert(result.head.getSeq(3).head.asInstanceOf[HashMap[String, _]]
       .keys.toSeq.sortWith(_ < _).mkString(",").equals(str2))
   }
